requirements.txt
requests==2.13.0
cassandra-driver
